Distance in Lead Group implemented

+++ app/src/main/java/ch/hsr/sa/radiotour/dataaccess/models/RiderStageConnectionComparatorDistanceInLead.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 public class RiderStageConnectionComparatorDistanceInLead implements Comparator<RiderStageConnection> {
     @Override
     public int compare(RiderStageConnection r1, RiderStageConnection r2) {
-        return Long.compare(r2.getDistanceInLeadGroup(), r1.getDistanceInLeadGroup());
+        return Double.compare(r2.getDistanceInLeadGroup(), r1.getDistanceInLeadGroup());
     }
 }
